xnet2figure/requirements.txt
python-igraph
matplotlib
numpy
